Use subsample in field analysis to reduce running time + other updates

+++ analyze_fields.py
@@ -1,10 +1,20 @@
+import os
 import json
 import pymongo
 import psycopg2
 from database_connectors import mongo_client, psql_cursor
 
-def combine_fields(documents, combined = {}):
+SAMPLE_SIZE = 5000000
+
+def combine_fields(documents, combined = {}, document_count = -1, collection_name = ""):
+        
+	document_index = 1
 	for document in documents:
+		if document_count >= 0:
+			with open("log.txt", 'w') as f:
+				f.write((collection_name + ": " if collection_name else "") + "Document %d of %d" % (document_index, document_count))
+			document_index += 1
+
 		for field in document:
 
 			value = document[field]
@@ -18,6 +28,7 @@ def combine_fields(documents, combined = {}):
 				while i < len(types):
 					if type(types[i]) == dict:
 						break
+					i += 1
 				if i == len(types):
 					types.append({})
 					
@@ -32,12 +43,14 @@ def dump_type(o):
     return str(o)
 
 def run():
-    db = mongo_client.ghtorrent_stripped
-    for name in db.list_collection_names():
-        print("Analyzing", name)
-        fields = combine_fields(db[name].find())
-        with open(name + ".json", 'w') as f:
-            json.dump(f, default=dump_type, indent=2)
+	db = mongo_client.ghtorrent_stripped
+	for name in db.list_collection_names():
+		if not os.path.exists(name + '.json'):
+			cursor = db[name].aggregate([ { "$sample": { "size": SAMPLE_SIZE } } ])
+			count = min(db[name].count_documents({}), SAMPLE_SIZE)
+			fields = combine_fields(cursor, document_count=count, collection_name=name)
+			with open(name + ".json", 'w') as f:
+				json.dump(fields, fp=f, default=dump_type, indent=2)
 
 if __name__ == "__main__":
     run()
